+++ src/Pages/HomePage.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import Carousal from '../customer/components/Carousal'
+
+const HomePage = () => {
+
+  return (
+    <div>
+      <Carousal />
+    </div>
+  )
+}
+
+export default HomePage
