--- src/customer/components/Carousal.js
@@ -1,38 +1,0 @@
-import React from 'react'
-import { Carousel } from "@material-tailwind/react";
-
-
-const Carousal = () => {
-    return (
-        <div className='big_carousal max-vh-100'>
-            <div id="carouselExampleIndicators" className="carousel slide " data-bs-ride="carousel">
-  <div className="carousel-indicators">
-    <button type="button" data-bs-target="#carouselExampleIndicators" data-bs-slide-to="0" className="active" aria-current="true" aria-label="Slide 1"></button>
-    <button type="button" data-bs-target="#carouselExampleIndicators" data-bs-slide-to="1" aria-label="Slide 2"></button>
-    <button type="button" data-bs-target="#carouselExampleIndicators" data-bs-slide-to="2" aria-label="Slide 3"></button>
-  </div>
-  <div className="carousel-inner">
-    <div className="carousel-item active" data-bs-interval="2000">
-      <img src="https://vedabyte.com/wp-content/uploads/2022/10/banner-1.jpg" className="d-block img-fluid rounded w-100" alt="..."/>
-    </div>
-    <div className="carousel-item" data-bs-interval="2000">
-      <img src="https://vedabyte.com/wp-content/uploads/2022/10/banner-1.jpg" className="d-block img-fluid rounded w-100" alt="..."/>
-    </div>
-    <div className="carousel-item" data-bs-interval="2000">
-      <img src="https://vedabyte.com/wp-content/uploads/2022/10/banner-1.jpg" className="d-block img-fluid rounded w-100" alt="..."/>
-    </div>
-  </div>
-  <button className="carousel-control-prev" type="button" data-bs-target="#carouselExampleIndicators" data-bs-slide="prev">
-    <span className="carousel-control-prev-icon" aria-hidden="true"></span>
-    <span className="visually-hidden">Previous</span>
-  </button>
-  <button className="carousel-control-next" type="button" data-bs-target="#carouselExampleIndicators" data-bs-slide="next">
-    <span className="carousel-control-next-icon" aria-hidden="true"></span>
-    <span className="visually-hidden">Next</span>
-  </button>
-</div>
-           </div>
-    )
-}
-
-export default Carousal
