--- src/customer/pages/HomePage.js
@@ -1,13 +1,0 @@
-import React from 'react'
-import Carousal from '../components/Carousal'
-
-const HomePage = () => {
-
-  return (
-    <div>
-     <Carousal/>
-    </div>
-  )
-}
-
-export default HomePage
